Move user functionality from engine to spec dummy app
Cut out all the code for users, Devise or CanCanCan and shift it to the dummy app we use for RSpec testing.

We also remove the specs for users and sessions, as the host applications should handle this.
