--- app/models/waste_carriers_engine/ability.rb
@@ -1,10 +1,0 @@
-module WasteCarriersEngine
-  class Ability
-    include CanCan::Ability
-
-    def initialize(user)
-      can :read, WasteCarriersEngine::Registration, account_email: user.email
-      can :manage, WasteCarriersEngine::TransientRegistration, account_email: user.email
-    end
-  end
-end
