+++ spec/dummy/app/models/ability.rb
@@ -1,0 +1,8 @@
+class Ability
+  include CanCan::Ability
+
+  def initialize(user)
+    can :read, WasteCarriersEngine::Registration, account_email: user.email
+    can :manage, WasteCarriersEngine::TransientRegistration, account_email: user.email
+  end
+end
