--- spec/models/waste_carriers_engine/user_spec.rb
@@ -1,55 +1,0 @@
-require "rails_helper"
-
-module WasteCarriersEngine
-  RSpec.describe User, type: :model do
-    describe "#password" do
-      context "when the user's password meets the requirements" do
-        let(:user) { build(:user, password: "Secret123") }
-
-        it "should be valid" do
-          expect(user).to be_valid
-        end
-      end
-
-      context "when the user's password is blank" do
-        let(:user) { build(:user, password: "") }
-
-        it "should not be valid" do
-          expect(user).to_not be_valid
-        end
-      end
-
-      context "when the user's password has no lowercase letters" do
-        let(:user) { build(:user, password: "SECRET123") }
-
-        it "should not be valid" do
-          expect(user).to_not be_valid
-        end
-      end
-
-      context "when the user's password has no uppercase letters" do
-        let(:user) { build(:user, password: "secret123") }
-
-        it "should not be valid" do
-          expect(user).to_not be_valid
-        end
-      end
-
-      context "when the user's password has no numbers" do
-        let(:user) { build(:user, password: "SuperSecret") }
-
-        it "should not be valid" do
-          expect(user).to_not be_valid
-        end
-      end
-
-      context "when the user's password is too short" do
-        let(:user) { build(:user, password: "Sec123") }
-
-        it "should not be valid" do
-          expect(user).to_not be_valid
-        end
-      end
-    end
-  end
-end
