+++ waste_carriers_engine.gemspec
@@ -30,13 +30,6 @@ Gem::Specification.new do |s|
   # Use AASM to manage states and transitions
   s.add_dependency "aasm", "~> 4.12"
 
-  # Use CanCanCan for user roles and permissions
-  # Version 2.0 doesn't support Mongoid, so we're locked to an earlier one
-  s.add_dependency "cancancan", "~> 1.10"
-
-  # Use Devise for user authentication
-  s.add_dependency "devise", ">= 4.4.3"
-
   # Use rest-client for external requests, eg. to Companies House
   s.add_dependency "rest-client", "~> 2.0"
